Modified the differentiable histogram function
Still skeptical if the differentiable histogram really works since the gradients it produces are always zero.

My feeling is the histogram is not differentiable at its core since it involves logical operations resulting in boolean values. This is probably also the reason `torch.histc` is also not differentiable  (https://discuss.pytorch.org/t/the-logical-or-and-not-function-about-pytorch/32718/5)